revert change to dgl sampler

+++ python/cugraph-dgl/cugraph_dgl/dataloading/neighbor_sampler.py
@@ -194,7 +194,7 @@ class NeighborSampler(Sampler):
 
         if g.is_homogeneous:
             indices = torch.concat(list(indices))
-            reader = ds.sample_from_nodes(indices, batch_size=batch_size)
+            reader = ds.sample_from_nodes(indices.long(), batch_size=batch_size)
             return HomogeneousSampleReader(reader, self.output_format, self.edge_dir)
 
         raise ValueError(
